+++ kicad_plugin/__init__.py
@@ -1,0 +1,3 @@
+"""PyAutoRoute KiCad action plugin."""
+
+from . import plugin